+++ runtime/runtime.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #include <stdio.h>
 #include <time.h>
 
-static inline void matrix_init(const long long value, const size_t rows, const size_t cols, long long m[rows][cols])
+static inline void matrix_init(const int64_t value, const size_t rows, const size_t cols, int64_t m[rows][cols])
 {
 	for (size_t row = 0; row < rows; ++row) {
 		for (size_t col = 0; col < cols; ++col) {
@@ -10,11 +11,11 @@ static inline void matrix_init(const long long value, const size_t rows, const s
 	}
 }
 
-void multiply(const size_t M, const size_t K, const size_t N ,const long long left[M][K], const long long right[K][N], long long result[M][N])
+void multiply(const size_t M, const size_t K, const size_t N ,const int64_t left[M][K], const int64_t right[K][N], int64_t result[M][N])
 {
 	for (size_t row = 0; row < M; ++row) {
 		for (size_t col = 0; col < N; ++col) {
-			long long sum = 0;
+			int64_t sum = 0;
 			for (size_t j = 0; j < K; ++j) {
 				sum += left[row][j] * right[j][col];
 			}
@@ -23,10 +24,10 @@ void multiply(const size_t M, const size_t K, const size_t N ,const long long le
 	}
 }
 
-void matrix_print(const size_t rows, const size_t cols, const long long m[rows][cols]) {
+void matrix_print(const size_t rows, const size_t cols, const int64_t m[rows][cols]) {
 	for (size_t row = 0; row < rows; ++row) {
 		for (size_t col = 0; col < cols; ++col) {
-			printf("%lld ", m[row][col]);
+			printf("%ld ", m[row][col]);
 		}
 		printf("\n");
 	}
@@ -41,16 +42,16 @@ static inline void print_time(struct timespec start, struct timespec end)
 
 
 int main() {
-	long long left[50][20];
+	int64_t left[50][20];
 	matrix_init(5, 50, 20, left);
 
-	long long right[20][15];
+	int64_t right[20][15];
 	matrix_init(2, 20, 15, right);
 
 	struct timespec start, end;
 	clock_gettime(CLOCK_PROCESS_CPUTIME_ID, &start);
 
-	long long result[50][15];
+	int64_t result[50][15];
 	multiply(50, 20, 15, left, right, result);
 
 	clock_gettime(CLOCK_PROCESS_CPUTIME_ID, &end);
